More fixes to issues noted by cppcheck

+++ examples/pipe/interlock_test.c
@@ -172,8 +172,8 @@ int interlock_test(void)
     }
   else if (ret != 0)
     {
-      fprintf(stderr, "interlock_test: Read %d bytes of data -- aborting: %d\n",
-              nbytes, errno);
+      fprintf(stderr, "interlock_test: Read %ld bytes of data -- aborting: %d\n",
+              (long)nbytes, errno);
       ret = 5;
       goto errout_with_file;
     }
